Make copy of id3 picture data before passing to user
There's a weird bug were concurrent writes of
the picture data from different mp3 files results
in one or more of the written picture data files to
be corrupt.

A quick fix for now is to make an internal copy of the
buffer data before passing on to the user. A deeper
investigation / analysis is required as it will
theoretically effect other container types.

Fixes #49