Don't automatically split values by the / character
It's impossible to deal with multiple values within a
single tag in id3v2.3 as no escape character is defined
for the / separator so we defer the decision on how to
deal with these values to the user. In id3v2.4 this isn't
an issue as 0x00 is defined for the separator.