Re-ordered fsize() so that a user-specified fileSize option will override autmatic fileSize determination.
When the user specifies a fileSize it's not a suggestion: it's the law. Consequently there's no need to go checking for the filesize, just run with it.  If it's wrong, that's the user's problem.

As to the comment change: IMHO code should be clean of all vulgarity - you never know who might want to work on the code and it's better to keep it clean.