feat: allow to populate common tags with information of all tag headers
Adding a new parsing option to enable a "best effort" approach while
trying to get information from all available tag headers when possible.

If the same information is available in multiple headers, the priority
is given to the newest version of the header.

eg:
[ID3v1.1] artist: "Some artist"
[ID3v2.4] artist: "Another artist"

The result will be that artist is "Another artist" in common.

Relates to: https://github.com/Borewit/music-metadata/issues/72